requirements.txt
Django==2.1.2
djangorestframework==3.9.0
zappa==0.47.0
uWSGI==2.0.17.1
requests==2.20.0
zappa-django-utils==0.4.0
